graylog2-server/src/main/java/org/graylog2/shared/rest/OptionalResponseFilter.java
/*
 * Copyright (C) 2020 Graylog, Inc.
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the Server Side Public License, version 1,
 * as published by MongoDB, Inc.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 * Server Side Public License for more details.
 *
 * You should have received a copy of the Server Side Public License
 * along with this program. If not, see
 * <http://www.mongodb.com/licensing/server-side-public-license>.
 */
package org.graylog2.shared.rest;

import jakarta.ws.rs.container.ContainerRequestContext;
import jakarta.ws.rs.container.ContainerResponseContext;
import jakarta.ws.rs.container.ContainerResponseFilter;
import jakarta.ws.rs.core.Response;

import java.util.Optional;

public class OptionalResponseFilter implements ContainerResponseFilter {

    @Override
    public void filter(final ContainerRequestContext requestContext,
                       ContainerResponseContext response) {
        final Object entity = response.getEntity();
        if (isEmptyOptional(entity)) {
            response.setStatus(Response.Status.NO_CONTENT.getStatusCode());
            response.setEntity(null);
        }
    }

    private boolean isEmptyOptional(final Object entity) {
        return entity instanceof Optional<?> && ((Optional<?>) entity).isEmpty();
    }
}
